// Angular12Crud/src/app/models/operation.model.ts
export class Operation {
    id?: any; 
    titreoperation?: string; 
    entreprise?: string; 
    annee_de_livraison?: string; 
    ville?: string;
    ppi?: string; 
    lycee?: string; 
    notification_du_marche?: string;
    codeuai?: string; 
    longitude?: any;
    etat_d_avancement?: string; 
    montant_des_ap_votes_en_meu?: any;
    cao_attribution?: string;
    latitude?: any;
    maitrise_d_oeuvre?: string; 
    mode_de_devolution?: string; 
    annee_d_individualisation?: string; 
    enveloppe_prev_en_meu?: any
}